requirements.txt

datasets
editdistance
kenlm
hydra-core
matplotlib
numpy
speechbrain>=0.5.12
tensorboard
torch>=1.11.0
torch_audiomentations
torchaudio>=0.11.0
torchmetrics
torchvision>=0.12.0
tqdm
pandas
pillow
pyctcdecode
pyloudnorm
wandb
wget